Created a prerenderer script which is transpiled using webpack Then called to actually ✨prerender✨ with global css and everything
