+++ src/app/app.node.module.ts
@@ -6,12 +6,13 @@ import { AppRoutingModule } from './app-routing.module';
 import { CoreModule } from './core/core.module';
 import { HelloModule } from './hello/hello.module';
 
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
-  moduleId: module.id,
   selector: 'rio-app',
   templateUrl: './app.component.html',
+  styleUrls: ['../styles/index.css'],
+  encapsulation: ViewEncapsulation.None,
 })
 export class AppComponent { }
