+++ src/app/hello/hello.component.ts
@@ -1,8 +1,17 @@
 import { Component } from '@angular/core';
 
 @Component({
-  moduleId: module.id,
   templateUrl: './hello.component.html',
   styleUrls: ['./hello.component.css'],
 })
-export class HelloComponent { }
+export class HelloComponent {
+  count: number = 1
+
+  increment() {
+    this.count++;
+  }
+
+  decrement() {
+    this.count--;
+  }
+}
