+++ webpack/loaders.js
@@ -49,7 +49,7 @@ exports.html = {
 
 exports.localCss = {
   test: /\.css$/,
-  include: path.resolve(process.cwd(), 'src', 'app'),
+  // include: path.resolve(process.cwd(), 'src', 'app'),
   use: [
     'to-string-loader',
     'css-loader?-minimize',
@@ -58,18 +58,18 @@ exports.localCss = {
   exclude: /node_modules/,
 };
 
-exports.globalCss = {
-  test: /\.css$/,
-  include: path.resolve(process.cwd(), 'src', 'styles'),
-  use: ExtractTextPlugin.extract({
-    fallback: 'style-loader',
-    use: [
-      'css-loader?-minimize',
-      'postcss-loader',
-    ],
-  }),
-  exclude: /node_modules/,
-};
+// exports.globalCss = {
+//   test: /\.css$/,
+//   include: path.resolve(process.cwd(), 'src', 'styles'),
+//   use: ExtractTextPlugin.extract({
+//     fallback: 'style-loader',
+//     use: [
+//       'css-loader?-minimize',
+//       'postcss-loader',
+//     ],
+//   }),
+//   exclude: /node_modules/,
+// };
 
 exports.svg = makeFileLoader(/\.svg$/);
 exports.eot = makeFileLoader(/\.eot$/);
